requirements.txt
asgiref==3.8.1
diff-match-patch==20230430
Django==5.0.6
django-cors-headers==4.4.0
django-filter==24.2
django-import-export==4.0.9
djangorestframework==3.15.2
et-xmlfile==1.1.0
openpyxl==3.1.4
sqlparse==0.5.0
tablib==3.5.0
tzdata==2024.1
